% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrTable.R
\name{corrTable}
\alias{corrTable}
\title{make correlation table}
\usage{
corrTable(xData, method = c("pearson"))
}
\arguments{
\item{method, }{pearson or}

\item{xdata, }{dataset}
}
\description{
make correlation table
}
\examples{

corrTable()
}
\keyword{correlation}
\keyword{table}
