% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp.R
\name{ggp}
\alias{ggp}
\title{ggp}
\usage{
ggp(y, x)
}
\arguments{
\item{y, }{y variable}

\item{x, }{x variable}
}
\description{
Function to plot y against x and use ggsave
}
\examples{
ggp(GDWLIMP,SMOOTH)
ggp()
}
\keyword{scatterplot}
