% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impyrtable.R
\name{impyrtable}
\alias{impyrtable}
\title{impyrtable}
\usage{
impyrtable(year, gem = TRUE)
}
\arguments{
\item{year, }{year to filter the data on}

\item{gem, }{wether to plot and save a table with mean impairment
(default=TRUE) sample table or to plot and save sample table}
}
\description{
Function to ouput sample table by industry in a given year,
or to output average impairment table by industry in a given year.
}
\examples{
impyrtable()
}
\keyword{impyrtable}
