% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsmaker.R
\name{modelsmaker}
\alias{modelsmaker}
\title{modelsmaker}
\usage{
modelsmaker(model, excl, data = dsCase)
}
\arguments{
\item{model}{The formula object representing your raw model}

\item{excl}{A variable to exclude in the new model}
}
\description{
Takes a formula object as input and outputs two new formulae.
It's use is to create the models required in residual analysis.
newmodelIV will be a formula with the original independent variable.
newmodelDV will be a formula with excl as independent variable
}
\examples{
modelsmaker()
}
\keyword{modelmaker}
