% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA.R
\name{RA}
\alias{RA}
\title{RA}
\usage{
RA(model)
}
\arguments{
\item{model, }{The formula object to analyse}
}
\description{
Function to automate residual analysis.
It outputs several partial regression plots
}
\examples{
RA()
}
\keyword{Residual-analysis}
