% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualAna.R
\name{resdiualAna}
\alias{resdiualAna}
\title{resdiualAna}
\usage{
resdiualAna(model)
}
\arguments{
\item{model, }{The formula object to analyse}
}
\description{
Function to automate residual analysis.
It outputs several partial regression plots
}
\examples{
resdiualAna()
}
\keyword{Residual-analysis}
