% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualAna.R
\name{residualAna}
\alias{residualAna}
\title{residualAna}
\usage{
residualAna(model, data = dsCase)
}
\arguments{
\item{model, }{The formula object to analyse}
}
\description{
Function to automate residual analysis.
It outputs several partial regression plots
}
\examples{
residualAna()
}
\keyword{Residual-analysis}
