% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionator.R
\name{regressionator}
\alias{regressionator}
\title{regressionator}
\usage{
regressionator(
  vars,
  data = dsCase,
  dep = vars[1],
  residualAna = FALSE,
  quadr = "",
  method = "OLS",
  name = ""
)
}
\arguments{
\item{vars}{List of variables to make formula, or formula object}

\item{dep}{Dependent variable, defaults to first item in vars}

\item{residualAna}{Wether to do a residualanalysis}

\item{quadr}{non-linear variables to add as quadratic}

\item{method}{OLS, logit, probit or tobit}

\item{name}{A name for stargazer output. .doc extension is added}

\item{dsCase}{Dataset of your case, defaults to, possibly non-existent, dsCase}
}
\description{
Create formula object from list of variables and outputs in stargazer
}
\examples{
regressionator(vars = c(k, EXTERNAL, BATH:SMOOTH))
regressionator()
}
\keyword{regressionator}
